Added buttons to load files incase users don't have one
Made it so the the the buttons and file inputs are made in the
JavaScript file instead of the HTML.  Added the MIDI files that the
buttons load.  Also made it so that the buttons and the file input use
the same function to load the file into PicoAudio.

Should probably make a function to make the html buttons that load the
file to reduce the amount of duplicate code.